Move source code into src, export in library file and update example.
